+++ docs/changes.rst
@@ -19,6 +19,26 @@ Change log
 ----------
 
 
+Version 0.5.0
+^^^^^^^^^^^^^
+
+Released: not yet
+
+**Incompatible changes:**
+
+**Deprecations:**
+
+**Bug fixes:**
+
+**Enhancements:**
+
+**Known issues:**
+
+* See `list of open issues`_.
+
+.. _`list of open issues`: https://github.com/zhmcclient/zhmc-ansible-modules/issues
+
+
 Version 0.4.0
 ^^^^^^^^^^^^^